Create photobatchTrenchReport
fileReport.py can generate a .csv for a folder containing "photobatch" subfolders. Generated report will give number of jpeg images, indicate which files are present or absent, and will give the average elevation of coded targets in a coded target text file (if properly formatted). On execution, the file prompts for the path containing photobatch folders.